# open_api_spec.yaml
openapi: 3.0.3
info:
  title: API de Tarjetas de Refrigerio
  description: Backend desarrollado en Rust (Actix-web) para la gestión de tarjetas, comprobantes y pagos.
  version: 1.0.0
servers:
  - url: http://89.117.63.107:9695/tarjetasRefrigerio
    description: Servidor de Pruebas
  - url: /tarjetasRefrigerio
    description: Servidor Local (Ruta relativa)
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Token JWT almacenado en una cookie HTTP-only.
  schemas:
    Response:
      type: object
      properties:
        codigo:
          type: integer
          example: 200
        status:
          type: string
          nullable: true
          example: "Exito"
        mensaje:
          type: string
          nullable: true
          example: "Operación realizada correctamente"
...
    ComprobanteNuevoReq:
      type: object
      required:
        - usuario
        - numero
        - valor
        - file
      properties:
        usuario:
          type: string
        numero:
          type: string
        valor:
          type: number
          format: float
        file:
          type: object
          properties:
            contenido:
              type: string
              description: Base64 del archivo
            extension:
              type: string
        representados:
          type: array
          items:
            type: integer
...
paths:
...
  /comprobante/guardar:
    post:
      tags:
        - Comprobantes
      summary: Registrar un nuevo comprobante
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComprobanteNuevoReq'
      responses:
        '200':
          description: Comprobante guardado físicamente y registrado en DB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /comprobante/autorizar:
    post:
      tags:
        - Comprobantes
      summary: Autorizar una lista de comprobantes
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                numeros:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Resumen de aprobación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
...